Deserializing candidates from SDP
This commit adds a deserialize candidate from SDP function and
incorporates it into the example program.

The lib.rs file is now really messy though, need to clean it up soon.
Also made the socket of Candidate an Option, so that remote candidates
don't need an actual socket... not a fan of how messy it's gotten.